Fix styled path check
