+++ .changeset/big-llamas-look.md
@@ -1,0 +1,5 @@
+---
+'@compiled/babel-plugin': patch
+---
+
+Fix styled path check
