+++ packages/babel-plugin/src/utils/ast.tsx
@@ -5,7 +5,7 @@ import { parse } from '@babel/parser';
 import fs from 'fs';
 import { dirname, join } from 'path';
 import resolve from 'resolve';
-import type { Metadata } from '../types';
+import type { Metadata, State } from '../types';
 import type { PartialBindingWithMeta } from './types';
 
 /**
@@ -40,76 +40,121 @@ export const getPathOfNode = <TNode extends unknown>(
 };
 
 /**
- * Returns `true` if the expression is using `css` from `@compiled/react` as a tagged template expression
+ * Returns `true` if the node is using `css` from `@compiled/react` as a call expression
  *
- * @param node {t.Expression} The expression that is being checked
- * @param meta {Metadata} Useful metadata that can be used during the transformation
- * @returns {boolean} Whether the node is a css usage from compiled
- */
-export const isCompiledCSSTemplateLiteral = (
-  node: t.Expression,
-  meta: Metadata
-): node is t.TaggedTemplateExpression => {
-  return (
-    t.isTaggedTemplateExpression(node) &&
-    t.isIdentifier(node.tag) &&
-    node.tag.name === meta.state.compiledImports?.css
-  );
-};
-
-/**
- * Returns `true` if the expression is using `css` from `@compiled/react` as an object call expression.
- *
- * @param node {t.Expression} The expression that is being checked
- * @param meta {Metadata} Useful metadata that can be used during the transformation
+ * @param node {t.Expression} The expression node that is being checked
+ * @param state {State} Plugin state
  * @returns {boolean} Whether the node is a css usage from compiled
  */
 export const isCompiledCSSCallExpression = (
   node: t.Expression,
-  meta: Metadata
-): node is t.CallExpression => {
-  return (
-    t.isCallExpression(node) &&
-    t.isIdentifier(node.callee) &&
-    node.callee.name === meta.state.compiledImports?.css
-  );
-};
+  state: State
+): node is t.CallExpression =>
+  t.isCallExpression(node) &&
+  t.isIdentifier(node.callee) &&
+  node.callee.name === state.compiledImports?.css;
 
 /**
- * Returns `true` if the expression is using `keyframes` from `@compiled/react` as a tagged template expression.
+ * Returns `true` if the node is using `css` from `@compiled/react` as a tagged template expression
  *
- * @param node {t.Node} The expression that is being checked
- * @param meta {Metadata} Useful metadata that can be used during the transformation
- * @returns {boolean} Whether the node is a compiled keyframe
+ * @param node {t.Expression} The expression node that is being checked
+ * @param state {State} Plugin state
+ * @returns {boolean} Whether the node is a css usage from compiled
  */
-export const isCompiledKeyframesTaggedTemplateExpression = (
-  node: t.Node,
-  meta: Metadata
-): node is t.TaggedTemplateExpression => {
-  return (
-    t.isTaggedTemplateExpression(node) &&
-    t.isIdentifier(node.tag) &&
-    node.tag.name === meta.state.compiledImports?.keyframes
-  );
-};
+export const isCompiledCSSTaggedTemplateExpression = (
+  node: t.Expression,
+  state: State
+): node is t.TaggedTemplateExpression =>
+  t.isTaggedTemplateExpression(node) &&
+  t.isIdentifier(node.tag) &&
+  node.tag.name === state.compiledImports?.css;
 
 /**
- * Returns `true` if the expression is using `keyframes` from `@compiled/react` as a call expression.
+ * Returns `true` if the node is using `keyframes` from `@compiled/react` as a call expression
  *
- * @param node {t.Node} The expression that is being checked
- * @param meta {Metadata} Useful metadata that can be used during the transformation
+ * @param node {t.Node} The node that is being checked
+ * @param state {State} Plugin state
  * @returns {boolean} Whether the node is a compiled keyframe
  */
 export const isCompiledKeyframesCallExpression = (
   node: t.Node,
-  meta: Metadata
-): node is t.CallExpression => {
-  return (
-    t.isCallExpression(node) &&
-    t.isIdentifier(node.callee) &&
-    node.callee.name === meta.state.compiledImports?.keyframes
-  );
-};
+  state: State
+): node is t.CallExpression =>
+  t.isCallExpression(node) &&
+  t.isIdentifier(node.callee) &&
+  node.callee.name === state.compiledImports?.keyframes;
+
+/**
+ * Returns `true` if the node is using `keyframes` from `@compiled/react` as a tagged template expression
+ *
+ * @param node {t.Node} The node that is being checked
+ * @param state {State} Plugin state
+ * @returns {boolean} Whether the node is a compiled keyframe
+ */
+export const isCompiledKeyframesTaggedTemplateExpression = (
+  node: t.Node,
+  state: State
+): node is t.TaggedTemplateExpression =>
+  t.isTaggedTemplateExpression(node) &&
+  t.isIdentifier(node.tag) &&
+  node.tag.name === state.compiledImports?.keyframes;
+
+/**
+ * Returns `true` if the node is using `styled` from `@compiled/react` from a styled.tag usage
+ *
+ * @param node {t.Node} The node that is being checked
+ * @param state {State} Plugin state
+ * @returns {boolean} Whether the node is a styled usage from compiled
+ */
+const isCompiledStyledMemberExpression = (node: t.Node, state: State): node is t.MemberExpression =>
+  t.isMemberExpression(node) &&
+  t.isIdentifier(node.object) &&
+  node.object.name === state.compiledImports?.styled;
+
+/**
+ * Returns `true` if the node is using `styled` from `@compiled/react` from a styled(Component) usage
+ *
+ * @param node {t.Node} The node that is being checked
+ * @param state {State} Plugin state
+ * @returns {boolean} Whether the node is a styled usage from compiled
+ */
+const isCompiledStyledCompositionCallExpression = (
+  node: t.Node,
+  state: State
+): node is t.CallExpression =>
+  t.isCallExpression(node) &&
+  t.isIdentifier(node.callee) &&
+  node.callee.name === state.compiledImports?.styled;
+
+/**
+ * Returns `true` if the node is using `styled` from `@compiled/react` as a call expression
+ *
+ * @param node {t.Expression} The expression node that is being checked
+ * @param state {State} Plugin state
+ * @returns {boolean} Whether the node is a styled usage from compiled
+ */
+export const isCompiledStyledCallExpression = (
+  node: t.Expression,
+  state: State
+): node is t.CallExpression =>
+  t.isCallExpression(node) &&
+  (isCompiledStyledMemberExpression(node.callee, state) ||
+    isCompiledStyledCompositionCallExpression(node.callee, state));
+
+/**
+ * Returns `true` if the node is using `styled` from `@compiled/react` as a tagged template expression
+ *
+ * @param node {t.Node} The node that is being checked
+ * @param state {State} Plugin state
+ * @returns {boolean} Whether the node is a styled usage from compiled
+ */
+export const isCompiledStyledTaggedTemplateExpression = (
+  node: t.Node,
+  state: State
+): node is t.TaggedTemplateExpression =>
+  t.isTaggedTemplateExpression(node) &&
+  (isCompiledStyledMemberExpression(node.tag, state) ||
+    isCompiledStyledCompositionCallExpression(node.tag, state));
 
 /**
  * Builds a code frame error from a passed in node.
